+++ include/gtirb_pprinter/PrettyPrinter.hpp
@@ -401,6 +401,8 @@ private:
   std::set<gtirb::Addr> functionLastBlock;
   gtirb::Addr programCounter;
 
+  std::optional<gtirb::Addr> CFIStartProc;
+
   template <typename BlockType>
   void printBlockImpl(std::ostream& os, BlockType& block);
 };