+++ src/gtirb_pprinter/ElfBinaryPrinter.cpp
@@ -135,6 +135,12 @@ ElfBinaryPrinter::buildCompilerArgs(std::string outputFilename,
       break;
     }
   }
+  // add -m32 for x86 binaries
+  for (gtirb::Module& module : ir.modules()) {
+    if (module.getISA() == gtirb::ISA::IA32) {
+      args.push_back("-m32");
+    }
+  }
 
   if (debug) {
     std::cout << "Compiler arguments: ";